core: refactored company balance services
