+++ library/CoreBundle/Resources/config/model/provider/company.yml
@@ -8,15 +8,10 @@ services:
     arguments:
       $logger: '@logger'
 
-  Ivoz\Provider\Domain\Service\Company\IncrementBalance:
-    public: true
-    arguments:
-      $logger: '@logger'
-
   Ivoz\Provider\Domain\Service\Company\AbstractBalanceOperation:
     abstract:  true
     arguments:
-      $entityPersister: '@Ivoz\Core\Infrastructure\Domain\Service\DoctrineEntityPersister'
+      $entityTools: '@Ivoz\Core\Application\Service\EntityTools'
       $logger: '@logger'
       $client: '@Ivoz\Core\Infrastructure\Domain\Service\Cgrates\CompanyBalanceService'
       $syncBalanceService: '@Ivoz\Provider\Domain\Service\Company\SyncBalances'
