+++ library/Ivoz/Provider/Domain/Service/Company/IncrementBalance.php
@@ -21,9 +21,4 @@ class IncrementBalance extends AbstractBalanceOperation
             $company
         );
     }
-
-    public function getLastError()
-    {
-        return $this->lastError;
-    }
 }
